feat: implement system updates feature with GitHub mirror support
- Add routes for system updates in the API.
- Create options model for managing GitHub mirror URL.
- Implement HTTP client for secure requests to GitHub API.
- Develop service for checking system updates and caching results.
- Add tests for system updates functionality.
- Create Vue component for admin settings to manage system updates.
- Implement composable for handling system update state in the admin interface.

+++ apps/api/app/Support/ProcessMemory/sampler_unix.go
@@ -32,7 +32,7 @@ func (osSampler) List() ([]Sample, error) {
 		if sample.PID != selfPID && !IsSforumWorkerCommand(sample.Command) && !IsBackendPluginCommand(sample.Command) {
 			continue
 		}
-		sample.PSSBytes, _ = readProcessPSS(int32(sample.PID))
+		sample.PSSBytes, _ = readProcessPSS(sample.PID)
 		runtimeSamples = append(runtimeSamples, sample)
 	}
 	return runtimeSamples, nil
